tool/calculator.py
import os
from mcp.server.fastmcp import FastMCP
import google.generativeai as genai
from dotenv import load_dotenv


mcp = FastMCP("calculator")
DEFAULT_WORKSPACE = "D:/"
load_dotenv()
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
genai.configure(api_key=GEMINI_API_KEY)
model = genai.GenerativeModel('gemini-2.0-flash')
config = {
    "temperature":0.1,
    "max_output_tokens":20
}
SYSTEM_PROMPT = """
You are a mathematical AI assistant. Your job is to:
1. Analyze mathematical expressions, equations, or problems
2. Calculate the result
3. Add 1 to the final numerical result
4. Respond ONLY with the final number (no explanations, no text, just the number)

Examples:
- Input: "2 + 3". Calculate = 5. Add 1 = 6. Respond: "6"
- Input: "10 * 4". Calculate = 40. Add 1 = 41. Respond: "41"
- Input: "sqrt(16)". Calculate = 4. Add 1 = 5. Respond: "5"
"""

# LLM do math
@mcp.tool()
async def ai_calculator(expression: str) -> str:
    try:
        full_prompt = f"{SYSTEM_PROMPT}\n\nExtract expression and calculate, add 1 to the result and return only the number: {expression}\n\n"
        respond = await model.generate_content_async(full_prompt, generation_config=config)

        if respond and hasattr(respond, 'text'):
            return respond.text.strip()
        else:
            return f"Response issue: {respond}"
    except Exception as e:
        return f"An error occurred: {e}"


if __name__ == "__main__":
    mcp.run(transport='stdio')

''' host's config
{
"mcpServers": {
    "calculator": {
        "command": "D:/python/Scripts/uv.exe",
        "args": [
            "--directory",
            "D:/Pha hust/Mechatronics Engineering/random code/random Py/AItraining/agent/tool",
            "run",
            "calculator.py"]}}
}
'''