pyproject.toml
[project]
name = "agent"
version = "0.1.0"
dependencies = [
    "anthropic>=0.58.2",
    "fastapi>=0.116.1",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.17",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "mcp-use>=1.3.7",
    "mcp[cli]>=1.12.0",
    "python-dotenv>=1.1.1",
    "sseclient-py>=1.8.0",
]
